done view generation

+++ foodb_objectbox_adapter/test/foodb_objectbox_adapter_test.dart
@@ -7,14 +7,26 @@ import 'package:foodb_objectbox_adapter/foodb_objectbox_adapter.dart';
 import 'package:foodb_objectbox_adapter/objectbox.g.dart';
 import 'package:path/path.dart';
 
+Future<ObjectBoxAdapter> getAdapter(String dbName,
+    {bool persist = false}) async {
+  var directory = join(Directory.current.path, 'temp/test-$dbName');
+  final dir = Directory(directory);
+  late Store store;
+  if (!persist) {
+    if (dir.existsSync()) dir.deleteSync(recursive: true);
+    addTearDown(() {
+      store.close();
+      if (dir.existsSync()) dir.deleteSync(recursive: true);
+    });
+  }
+  store = await openStore(directory: directory);
+  return ObjectBoxAdapter(store);
+}
+
 class ObjectBoxTestContext extends FoodbTestContext {
   @override
   Future<Foodb> db(String dbName) async {
-    var directory = join(Directory.current.path, 'temp/$dbName');
-    final dir = Directory(directory);
-    if (dir.existsSync()) dir.deleteSync(recursive: true);
-    var db = ObjectBoxAdapter(await openStore(directory: directory));
-    await db.destroy();
+    var db = await getAdapter(dbName);
     return Foodb.keyvalue(dbName: dbName, keyValueDb: db);
   }
 }
@@ -22,7 +34,7 @@ class ObjectBoxTestContext extends FoodbTestContext {
 void main() {
   final objectBox = ObjectBoxTestContext();
   final couchdb = CouchdbTestContext();
-  replicateBenchmarkTest(1000, 3, objectBox);
+  // replicateBenchmarkTest(1000, 3, objectBox);
   // group('couchdb > objectBox', () {
   //   replicateTest().forEach((t) {
   //     t(couchdb, objectBox);
@@ -33,9 +45,9 @@ void main() {
   //     t(objectBox, couchdb);
   //   });
   // });
-  // allDocTest().forEach((fn) {
-  //   fn(objectBox);
-  // });
+  allDocTest().forEach((fn) {
+    fn(objectBox);
+  });
   // getTest().forEach((fn) {
   //   fn(objectBox);
   // });
